fix: rename zapsource to _try_source
Since this function is meant to be used locally we can prefix it with
underscore to hint that this function shouldn't be used externally. Also
the function doesn't return any status value which makes the `||`
unnecessary and since we check that the file exist before sourcing it's
not a problem to `_try_source` a non-existing file.